+++ custom_components/idleon/const.py
@@ -8,7 +8,7 @@ from homeassistant.const import Platform
 
 DOMAIN = "idleon"
 NAME = "Legends of Idleon"
-VERSION = "0.16.0"
+VERSION = "0.17.0"
 
 PLATFORMS = (Platform.SENSOR, Platform.BINARY_SENSOR)